fix(Android, FormSheet): Create SheetAnimationCoordinator to synchronize animations coming from different sources (#3914)
## Description

FormSheet with `fitToContents` and `SafeAreaView` applies window insets
during the entering transition, which triggers an extra `onLayout` pass
on the ContentWrapper. This resulted in starting a concurrent
ValueAnimator on translationY - conflicting with the slide-in animator
already running, causing a race between them and wrong sheet
positioning.

I'm introducing `SheetAnimationCoordinator`, which will be a single
source of truth for the `FormSheet's` transitions. I'm adding a guard in
`updateSheetContentHeightWithAnimation` - if an entry/exit animation is
already running, the height performs an immediate metrics update and
layout without touching translationY. The slide-in animator should have
a priority over the sheet height change animation.

> [!WARNING]
> Fit to contents (TextInput) case from Test3336 is still not working -
handling keyboard inset has its own animation which I'll synchronize in
the followup PR. Ticket:
https://github.com/software-mansion/react-native-screens-labs/issues/1170

Closes:
https://github.com/software-mansion/react-native-screens-labs/issues/968

## Changes

- introduced `SheetAnimationCoordinator`, which should own all
BottomSheet transitions
- replaced enter/exit animation and keyboard handling in `SheetDelegate`
with the delegate methods to the coordiantor
- moved `Screen` animations to `SheetAnimationCoordinator`
- moved `resolveClampedHeight` to SheetUtils - needs to be accessible
for both the screen and the coordinator.

## Before & after - visual documentation

| Before | After |
| --- | --- |
| <video
src="https://github.com/user-attachments/assets/5fcf64ee-cf33-41d1-ad1e-80c8cde31db0"
/> | <video
src="https://github.com/user-attachments/assets/495ad0b9-0939-4f2f-ae9a-b208bded147b"
/> |

## Test plan

Performed regression testing on Test3336 and TestFormSheet.

## Checklist

- [ ] Included code example that can be used to test this change.
- [ ] For visual changes, included screenshots / GIFs / recordings
documenting the change.
- [ ] For API changes, updated relevant public types.
- [ ] Ensured that CI passes

(cherry picked from commit 001ddea912dc93b3fac1874187aad74db82b0a5c)
